+++ crates/perry-runtime/src/node_submodules/mod.rs
@@ -22,17 +22,17 @@
 
 use std::cell::RefCell;
 use std::os::raw::c_int;
-use std::sync::atomic::{AtomicI64, AtomicPtr, Ordering};
+use std::sync::atomic::{AtomicI64, Ordering};
 
 use crate::closure::{
     js_closure_alloc, js_closure_call0, js_closure_call1, js_closure_call2, js_closure_call_array,
-    js_closure_get_capture_f64, js_closure_get_capture_ptr, js_closure_set_capture_f64,
-    js_closure_set_capture_ptr, js_register_closure_arity, ClosureHeader,
+    js_closure_get_capture_ptr, js_closure_set_capture_ptr, js_register_closure_arity,
+    ClosureHeader,
 };
 use crate::object::{
     js_object_alloc, js_object_get_field_by_name_f64, js_object_set_field_by_name, ObjectHeader,
 };
-use crate::string::{js_string_from_bytes, StringHeader};
+use crate::string::js_string_from_bytes;
 use crate::value::JSValue;
 
 mod diagnostics;
@@ -90,1294 +90,54 @@ struct SubmoduleSpec {
 
 macro_rules! thunk {
     ($name:ident, $msg:expr) => {
-        extern "C" fn $name(_closure: *const ClosureHeader, _arg: f64) -> f64 {
+        pub(crate) extern "C" fn $name(
+            _closure: *const crate::closure::ClosureHeader,
+            _arg: f64,
+        ) -> f64 {
             let msg: &'static str = $msg;
             let bytes = msg.as_bytes();
-            let header = js_string_from_bytes(bytes.as_ptr(), bytes.len() as u32);
+            let header = crate::string::js_string_from_bytes(bytes.as_ptr(), bytes.len() as u32);
             let err = crate::error::js_error_new_with_message(header);
-            let bits = JSValue::pointer(err as *const u8).bits();
+            let bits = crate::value::JSValue::pointer(err as *const u8).bits();
             crate::exception::js_throw(f64::from_bits(bits))
         }
     };
 }
 
-/// node:timers/promises.setTimeout(delay, value?) — a Promise that resolves
-/// with `value` (or undefined) after `delay` ms. Composes the existing
-/// promise-returning timer primitive; the closure dispatch pads a missing
-/// `value` arg with undefined (arity registered in `ensure_export_singleton`).
-/// Refs #1213.
-extern "C" fn timers_promises_set_timeout(
-    _closure: *const ClosureHeader,
-    delay_ms: f64,
-    value: f64,
-) -> f64 {
-    let promise = crate::timer::js_set_timeout_value(delay_ms, value);
-    crate::value::js_nanbox_pointer(promise as i64)
-}
-
-/// node:timers/promises.setImmediate(value?) — a Promise that resolves with
-/// `value` (or undefined) on a later turn. Refs #1213.
-extern "C" fn timers_promises_set_immediate(_closure: *const ClosureHeader, value: f64) -> f64 {
-    let promise = crate::timer::js_set_timeout_value(0.0, value);
-    crate::value::js_nanbox_pointer(promise as i64)
-}
-
-// ── node:timers namespace (`import * as timers from "node:timers"`) ──────────
-// Route to the SAME global timer runtime fns the bare globals use, so
-// `timers.setTimeout(...)` matches `setTimeout(...)`. NOTE: named imports
-// (`import { setTimeout } from "node:timers"`) deliberately bypass this and
-// keep the codegen global fast-path (which handles `setTimeout(fn, delay,
-// ...args)` varargs) — compile.rs skips registering node:timers named imports
-// as submodule exports. Refs #1213.
-fn callback_arg_to_i64(v: f64) -> i64 {
-    (v.to_bits() & 0x0000_FFFF_FFFF_FFFF) as i64
-}
-extern "C" fn timers_ns_set_timeout(_c: *const ClosureHeader, cb: f64, ms: f64) -> f64 {
-    crate::value::js_nanbox_pointer(crate::timer::js_set_timeout_callback(
-        callback_arg_to_i64(cb),
-        ms,
-    ))
-}
-extern "C" fn timers_ns_set_interval(_c: *const ClosureHeader, cb: f64, ms: f64) -> f64 {
-    crate::value::js_nanbox_pointer(crate::timer::setInterval(callback_arg_to_i64(cb), ms))
-}
-extern "C" fn timers_ns_set_immediate(_c: *const ClosureHeader, cb: f64) -> f64 {
-    crate::value::js_nanbox_pointer(crate::timer::js_set_immediate_callback(
-        callback_arg_to_i64(cb),
-    ))
-}
-extern "C" fn timers_ns_clear_timeout(_c: *const ClosureHeader, arg: f64) -> f64 {
-    crate::timer::js_clear_timeout_value(arg);
-    f64::from_bits(TAG_UNDEFINED)
-}
-extern "C" fn timers_ns_clear_interval(_c: *const ClosureHeader, arg: f64) -> f64 {
-    crate::timer::js_clear_interval_value(arg);
-    f64::from_bits(TAG_UNDEFINED)
-}
-// Immediates live in the shared timer pool; clearTimeout retains-out both pools.
-extern "C" fn timers_ns_clear_immediate(_c: *const ClosureHeader, arg: f64) -> f64 {
-    crate::timer::js_clear_timeout_value(arg);
-    f64::from_bits(TAG_UNDEFINED)
-}
-
-thunk!(
-    thunk_timers_setInterval,
-    "node:timers/promises.setInterval is not yet implemented in Perry (tracked by issue #793)."
-);
-thunk!(
-    thunk_timers_scheduler,
-    "node:timers/promises.scheduler is not yet implemented in Perry (tracked by issue #793)."
-);
-
-fn promise_value(value: f64) -> f64 {
-    let promise = crate::promise::js_promise_new();
-    crate::promise::js_promise_resolve(promise, value);
-    f64::from_bits(JSValue::pointer(promise as *const u8).bits())
-}
-
-fn promise_rejected(reason: f64) -> f64 {
-    let promise = crate::promise::js_promise_rejected(reason);
-    f64::from_bits(JSValue::pointer(promise as *const u8).bits())
-}
-
-fn promise_undefined() -> f64 {
-    promise_value(f64::from_bits(crate::value::TAG_UNDEFINED))
-}
-
-extern "C" fn thunk_fs_promises_readFile(
-    _closure: *const ClosureHeader,
-    path: f64,
-    encoding: f64,
-) -> f64 {
-    promise_value(crate::fs::js_fs_read_file_dispatch(path, encoding))
-}
-
-extern "C" fn thunk_fs_promises_open(
-    _closure: *const ClosureHeader,
-    path: f64,
-    flags: f64,
-    _mode: f64,
-) -> f64 {
-    // Probe before opening so a missing path rejects the Promise instead of
-    // resolving with a FileHandle whose `fd === -1`. Matches Node's behavior
-    // for `fs/promises.open(path)` on ENOENT/EACCES.
-    if let Some(err_val) = unsafe { crate::fs::fs_promises_open_probe_error(path, flags) } {
-        let promise = crate::promise::js_promise_rejected(err_val);
-        return f64::from_bits(JSValue::pointer(promise as *const u8).bits());
-    }
-    promise_value(crate::fs::js_fs_filehandle_open(path, flags))
-}
-
-extern "C" fn thunk_fs_promises_writeFile(
-    _closure: *const ClosureHeader,
-    path: f64,
-    data: f64,
-    options: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_write_file_sync_options(path, data, options);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_appendFile(
-    _closure: *const ClosureHeader,
-    path: f64,
-    data: f64,
-    options: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_append_file_sync_options(path, data, options);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_chmod(_closure: *const ClosureHeader, path: f64, mode: f64) -> f64 {
-    let _ = crate::fs::js_fs_chmod_sync(path, mode);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_chown(
-    _closure: *const ClosureHeader,
-    path: f64,
-    uid: f64,
-    gid: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_chown_sync(path, uid, gid);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_lchown(
-    _closure: *const ClosureHeader,
-    path: f64,
-    uid: f64,
-    gid: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_lchown_sync(path, uid, gid);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_mkdir(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_mkdir_sync_options(path, options);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_readdir(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    let raw = crate::fs::js_fs_readdir_sync(path, options);
-    promise_value(f64::from_bits(
-        JSValue::pointer(raw.to_bits() as *const u8).bits(),
-    ))
-}
-
-extern "C" fn thunk_fs_promises_stat(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    promise_value(crate::fs::js_fs_stat_sync_options(path, options))
-}
-
-extern "C" fn thunk_fs_promises_statfs(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    promise_value(crate::fs::js_fs_statfs_sync_options(path, options))
-}
-
-extern "C" fn thunk_fs_promises_lstat(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    promise_value(crate::fs::js_fs_lstat_sync_options(path, options))
-}
-
-extern "C" fn thunk_fs_promises_rm(_closure: *const ClosureHeader, path: f64, options: f64) -> f64 {
-    let _ = crate::fs::js_fs_rm_recursive_options(path, options);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_rmdir(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_rmdir_sync_options(path, options);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_unlink(_closure: *const ClosureHeader, path: f64) -> f64 {
-    let _ = crate::fs::js_fs_unlink_sync(path);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_rename(_closure: *const ClosureHeader, from: f64, to: f64) -> f64 {
-    let _ = crate::fs::js_fs_rename_sync(from, to);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_copyFile(
-    _closure: *const ClosureHeader,
-    from: f64,
-    to: f64,
-    flags: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_copy_file_sync_flags(from, to, flags);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_cp(
-    _closure: *const ClosureHeader,
-    from: f64,
-    to: f64,
-    options: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_cp_sync_options(from, to, options);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_truncate(
-    _closure: *const ClosureHeader,
-    path: f64,
-    len: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_truncate_sync(path, len);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_utimes(
-    _closure: *const ClosureHeader,
-    path: f64,
-    atime: f64,
-    mtime: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_utimes_sync(path, atime, mtime);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_lutimes(
-    _closure: *const ClosureHeader,
-    path: f64,
-    atime: f64,
-    mtime: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_lutimes_sync(path, atime, mtime);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_link(_closure: *const ClosureHeader, from: f64, to: f64) -> f64 {
-    let _ = crate::fs::js_fs_link_sync(from, to);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_symlink(
-    _closure: *const ClosureHeader,
-    target: f64,
-    path: f64,
-    _type: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_symlink_sync(target, path);
-    promise_undefined()
-}
-
-extern "C" fn thunk_fs_promises_readlink(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    promise_value(crate::fs::js_fs_readlink_dispatch(path, options))
-}
-
-extern "C" fn thunk_fs_promises_realpath(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    promise_value(crate::fs::js_fs_realpath_dispatch(path, options))
-}
-
-extern "C" fn thunk_fs_promises_mkdtemp(
-    _closure: *const ClosureHeader,
-    prefix: f64,
-    options: f64,
-) -> f64 {
-    promise_value(crate::fs::js_fs_mkdtemp_dispatch(prefix, options))
-}
-
-extern "C" fn thunk_fs_promises_opendir(_closure: *const ClosureHeader, path: f64) -> f64 {
-    promise_value(crate::fs::js_fs_opendir_sync(path))
-}
-
-extern "C" fn thunk_fs_promises_glob(
-    _closure: *const ClosureHeader,
-    pattern: f64,
-    options: f64,
-) -> f64 {
-    let raw = crate::fs::js_fs_glob_sync_options(pattern, options);
-    promise_value(f64::from_bits(
-        JSValue::pointer(raw.to_bits() as *const u8).bits(),
-    ))
-}
-
-extern "C" fn thunk_fs_promises_watch(
-    _closure: *const ClosureHeader,
-    path: f64,
-    options: f64,
-) -> f64 {
-    crate::fs::js_fs_watch(path, options, f64::from_bits(crate::value::TAG_UNDEFINED))
-}
-
-extern "C" fn thunk_fs_promises_access(
-    _closure: *const ClosureHeader,
-    path: f64,
-    mode: f64,
-) -> f64 {
-    let _ = crate::fs::js_fs_access_sync_mode(path, mode);
-    promise_undefined()
-}
-
-thunk!(thunk_readline_createInterface, "node:readline/promises.createInterface is not yet implemented in Perry (tracked by issue #793).");
-thunk!(
-    thunk_readline_Interface,
-    "node:readline/promises.Interface is not yet implemented in Perry (tracked by issue #793)."
-);
-thunk!(
-    thunk_readline_Readline,
-    "node:readline/promises.Readline is not yet implemented in Perry (tracked by issue #793)."
-);
-
-#[inline]
-fn undefined_value() -> f64 {
-    f64::from_bits(crate::value::TAG_UNDEFINED)
-}
-
-#[inline]
-fn value_from_ptr(ptr: *const u8) -> f64 {
-    f64::from_bits(JSValue::pointer(ptr).bits())
-}
-
-#[inline]
-fn raw_ptr_from_value(value: f64) -> usize {
-    let bits = value.to_bits();
-    let jsval = JSValue::from_bits(bits);
-    if jsval.is_pointer() || jsval.is_string() || jsval.is_bigint() {
-        return (bits & crate::value::POINTER_MASK) as usize;
-    }
-    if bits != 0 && bits < 0x0001_0000_0000_0000 {
-        return bits as usize;
-    }
-    0
-}
-
-#[inline]
-unsafe fn gc_type_for_ptr(raw: usize) -> Option<u8> {
-    if raw < crate::gc::GC_HEADER_SIZE + 0x1000 {
-        return None;
-    }
-    let header = (raw as *const u8).sub(crate::gc::GC_HEADER_SIZE) as *const crate::gc::GcHeader;
-    let gc_type = (*header).obj_type;
-    if gc_type <= crate::gc::GC_TYPE_MAX {
-        Some(gc_type)
-    } else {
-        None
-    }
-}
-
-fn object_ptr_from_value(value: f64) -> Option<*mut ObjectHeader> {
-    let raw = raw_ptr_from_value(value);
-    if raw < 0x10000 || crate::buffer::is_registered_buffer(raw) {
-        return None;
-    }
-    unsafe {
-        if gc_type_for_ptr(raw) != Some(crate::gc::GC_TYPE_OBJECT) {
-            return None;
-        }
-    }
-    Some(raw as *mut ObjectHeader)
-}
-
-fn get_object_property(value: f64, name: &[u8]) -> Option<f64> {
-    let obj = object_ptr_from_value(value)?;
-    let key = js_string_from_bytes(name.as_ptr(), name.len() as u32);
-    let value = js_object_get_field_by_name_f64(obj as *const ObjectHeader, key);
-    if JSValue::from_bits(value.to_bits()).is_undefined() {
-        None
-    } else {
-        Some(value)
-    }
-}
-
-fn options_signal(options: f64) -> Option<f64> {
-    let jsval = JSValue::from_bits(options.to_bits());
-    if jsval.is_undefined() || jsval.is_null() {
-        return None;
-    }
-    get_object_property(options, b"signal")
-}
-
-fn signal_aborted(signal: f64) -> bool {
-    get_object_property(signal, b"aborted").is_some_and(|v| crate::value::js_is_truthy(v) != 0)
-}
-
-fn abort_error_value() -> f64 {
-    let msg = b"AbortError";
-    let header = js_string_from_bytes(msg.as_ptr(), msg.len() as u32);
-    let err = crate::error::js_error_new_with_message(header);
-    value_from_ptr(err as *const u8)
-}
-
-fn signal_reason(signal: f64) -> f64 {
-    match get_object_property(signal, b"reason") {
-        Some(reason) if !JSValue::from_bits(reason.to_bits()).is_undefined() => reason,
-        _ => abort_error_value(),
-    }
-}
-
-extern "C" fn stream_promises_abort_listener(closure: *const ClosureHeader) -> f64 {
-    let promise_value = js_closure_get_capture_f64(closure, 0);
-    let signal = js_closure_get_capture_f64(closure, 1);
-    let promise =
-        crate::value::js_nanbox_get_pointer(promise_value) as *mut crate::promise::Promise;
-    crate::promise::js_promise_reject(promise, signal_reason(signal));
-    undefined_value()
-}
-
-fn promise_value_from_ptr(promise: *mut crate::promise::Promise) -> f64 {
-    value_from_ptr(promise as *const u8)
-}
-
-fn register_abort_listener(signal: f64, promise: *mut crate::promise::Promise) {
-    let Some(signal_obj) = object_ptr_from_value(signal) else {
-        return;
-    };
-    let closure = js_closure_alloc(stream_promises_abort_listener as *const u8, 2);
-    js_closure_set_capture_f64(closure, 0, promise_value_from_ptr(promise));
-    js_closure_set_capture_f64(closure, 1, signal);
-    let event = b"abort";
-    let event_str = js_string_from_bytes(event.as_ptr(), event.len() as u32);
-    let event_value = f64::from_bits(JSValue::string_ptr(event_str).bits());
-    let listener_value = value_from_ptr(closure as *const u8);
-    crate::url::js_abort_signal_add_listener(signal_obj, event_value, listener_value);
-}
-
-fn pending_abortable_promise(signal: f64) -> f64 {
-    let promise = crate::promise::js_promise_new();
-    register_abort_listener(signal, promise);
-    promise_value_from_ptr(promise)
-}
-
-fn invoke_destination_method(destination: f64, method: &[u8], args: &[f64]) -> f64 {
-    let Some(func) = get_object_property(destination, method) else {
-        return undefined_value();
-    };
-    let prev_this = crate::object::js_implicit_this_set(destination);
-    let result = unsafe { crate::closure::js_native_call_value(func, args.as_ptr(), args.len()) };
-    crate::object::js_implicit_this_set(prev_this);
-    result
-}
-
-fn write_chunks_to_destination(destination: f64, chunks: &[f64]) {
-    let undef = undefined_value();
-    for chunk in chunks {
-        let args = [*chunk, undef];
-        let _ = invoke_destination_method(destination, b"write", &args);
-    }
-    let end_args = [undef];
-    let _ = invoke_destination_method(destination, b"end", &end_args);
-}
-
-extern "C" fn thunk_streamP_pipeline(
-    _closure: *const ClosureHeader,
-    source: f64,
-    destination: f64,
-    options: f64,
-) -> f64 {
-    let signal = options_signal(options);
-    if let Some(signal) = signal {
-        if signal_aborted(signal) {
-            return promise_rejected(signal_reason(signal));
-        }
-    }
-
-    match crate::node_stream::js_node_stream_readable_chunks_result(source) {
-        Err(err) => promise_rejected(err),
-        Ok(Some(chunks)) => {
-            write_chunks_to_destination(destination, &chunks);
-            if let Some(signal) = signal {
-                if signal_aborted(signal) {
-                    return promise_rejected(signal_reason(signal));
-                }
-            }
-            promise_undefined()
-        }
-        Ok(None) => {
-            if let Some(signal) = signal {
-                pending_abortable_promise(signal)
-            } else if let Some(err) =
-                crate::node_stream::js_node_stream_hidden_error_after_read(source)
-            {
-                promise_rejected(err)
-            } else {
-                promise_undefined()
-            }
-        }
-    }
-}
-
-extern "C" fn thunk_streamP_finished(
-    _closure: *const ClosureHeader,
-    stream: f64,
-    options: f64,
-) -> f64 {
-    if let Some(signal) = options_signal(options) {
-        if signal_aborted(signal) {
-            return promise_rejected(signal_reason(signal));
-        }
-        if let Some(err) = crate::node_stream::js_node_stream_hidden_error_after_read(stream) {
-            return promise_rejected(err);
-        }
-        if crate::node_stream::js_node_stream_is_stub_ended_after_read(stream) {
-            return promise_undefined();
-        }
-        return pending_abortable_promise(signal);
-    }
-
-    if let Some(err) = crate::node_stream::js_node_stream_hidden_error_after_read(stream) {
-        promise_rejected(err)
-    } else {
-        promise_undefined()
-    }
-}
-
-fn buffer_from_bytes(
-    bytes: &[u8],
-    mark_array_buffer: bool,
-    mark_uint8_array: bool,
-) -> *mut crate::buffer::BufferHeader {
-    let buf = crate::buffer::buffer_alloc(bytes.len() as u32);
-    unsafe {
-        (*buf).length = bytes.len() as u32;
-        if !bytes.is_empty() {
-            std::ptr::copy_nonoverlapping(
-                bytes.as_ptr(),
-                crate::buffer::buffer_data_mut(buf),
-                bytes.len(),
-            );
-        }
-    }
-    if mark_array_buffer {
-        crate::buffer::mark_as_array_buffer(buf as usize);
-    }
-    if mark_uint8_array {
-        crate::buffer::mark_as_uint8array(buf as usize);
-    }
-    buf
-}
-
-fn bytes_to_buffer_value(bytes: &[u8]) -> f64 {
-    let buf = buffer_from_bytes(bytes, false, false);
-    f64::from_bits(JSValue::pointer(buf as *const u8).bits())
-}
-
-fn bytes_to_array_buffer_value(bytes: &[u8]) -> f64 {
-    let buf = buffer_from_bytes(bytes, true, false);
-    f64::from_bits(JSValue::pointer(buf as *const u8).bits())
-}
-
-fn bytes_to_uint8_array_value(bytes: &[u8]) -> f64 {
-    let buf = buffer_from_bytes(bytes, false, true);
-    f64::from_bits(JSValue::pointer(buf as *const u8).bits())
-}
-
-fn bytes_to_text_value(bytes: &[u8]) -> f64 {
-    let cow = String::from_utf8_lossy(bytes);
-    let ptr = js_string_from_bytes(cow.as_bytes().as_ptr(), cow.len() as u32);
-    f64::from_bits(JSValue::string_ptr(ptr).bits())
-}
-
-#[derive(Clone, Copy)]
-enum ConsumerKind {
-    Text = 0,
-    Json = 1,
-    Buffer = 2,
-    ArrayBuffer = 3,
-    Bytes = 4,
-    Blob = 5,
-}
-
-impl ConsumerKind {
-    fn from_i64(value: i64) -> Self {
-        match value {
-            1 => Self::Json,
-            2 => Self::Buffer,
-            3 => Self::ArrayBuffer,
-            4 => Self::Bytes,
-            5 => Self::Blob,
-            _ => Self::Text,
-        }
-    }
-
-    fn chunk_mode(self) -> ChunkMode {
-        match self {
-            Self::Text | Self::Json => ChunkMode::Text,
-            Self::Buffer | Self::ArrayBuffer | Self::Bytes | Self::Blob => ChunkMode::Binary,
-        }
-    }
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum ChunkMode {
-    Binary,
-    Text,
-}
-
-#[derive(Clone, Copy)]
-enum CollectMethod {
-    Next = 0,
-    Read = 1,
-}
-
-type StreamGetReaderFn = unsafe extern "C" fn(f64) -> f64;
-type StreamReaderReadFn = unsafe extern "C" fn(f64) -> *mut crate::Promise;
-
-static STREAM_CONSUMER_GET_READER: AtomicPtr<()> = AtomicPtr::new(std::ptr::null_mut());
-static STREAM_CONSUMER_READER_READ: AtomicPtr<()> = AtomicPtr::new(std::ptr::null_mut());
-
-#[no_mangle]
-pub extern "C" fn js_register_stream_consumer_callbacks(
-    get_reader: StreamGetReaderFn,
-    reader_read: StreamReaderReadFn,
-) {
-    STREAM_CONSUMER_GET_READER.store(get_reader as *mut (), Ordering::Release);
-    STREAM_CONSUMER_READER_READ.store(reader_read as *mut (), Ordering::Release);
-}
-
-impl CollectMethod {
-    fn from_i64(value: i64) -> Self {
-        if value == 1 {
-            Self::Read
-        } else {
-            Self::Next
-        }
-    }
-
-    fn name(self) -> &'static [u8] {
-        match self {
-            Self::Next => b"next",
-            Self::Read => b"read",
-        }
-    }
-}
-
-fn boxed_pointer(ptr: *const u8) -> f64 {
-    f64::from_bits(JSValue::pointer(ptr).bits())
-}
-
-fn is_integral_handle_value(value: f64) -> bool {
-    let jsval = JSValue::from_bits(value.to_bits());
-    (jsval.is_int32() && jsval.as_int32() > 0)
-        || (jsval.is_number() && value.is_finite() && value > 0.0 && value.fract() == 0.0)
-}
-
-fn is_undefined_value(value: f64) -> bool {
-    value.to_bits() == crate::value::TAG_UNDEFINED
-        || JSValue::from_bits(value.to_bits()).is_undefined()
-}
-
-fn raw_ptr_from_value(value: f64) -> usize {
-    let bits = value.to_bits();
-    let jsval = JSValue::from_bits(bits);
-    if jsval.is_pointer() || jsval.is_string() || jsval.is_bigint() {
-        return (bits & crate::value::POINTER_MASK) as usize;
-    }
-    if bits != 0 && bits < 0x0001_0000_0000_0000 {
-        return bits as usize;
-    }
-    0
-}
-
-unsafe fn gc_type_for_ptr(raw: usize) -> Option<u8> {
-    if raw < crate::gc::GC_HEADER_SIZE + 0x1000 {
-        return None;
-    }
-    let header = (raw as *const u8).sub(crate::gc::GC_HEADER_SIZE) as *const crate::gc::GcHeader;
-    let gc_type = (*header).obj_type;
-    if gc_type <= crate::gc::GC_TYPE_MAX {
-        Some(gc_type)
-    } else {
-        None
-    }
-}
-
-fn object_ptr_from_value(value: f64) -> Option<*const ObjectHeader> {
-    let raw = raw_ptr_from_value(value);
-    if raw < 0x10000 || crate::buffer::is_registered_buffer(raw) {
-        return None;
-    }
-    unsafe {
-        if gc_type_for_ptr(raw) != Some(crate::gc::GC_TYPE_OBJECT) {
-            return None;
-        }
-    }
-    Some(raw as *const ObjectHeader)
-}
-
-fn named_key(bytes: &[u8]) -> *const StringHeader {
-    js_string_from_bytes(bytes.as_ptr(), bytes.len() as u32)
-}
-
-fn get_named_value(value: f64, name: &[u8]) -> f64 {
-    let Some(obj) = object_ptr_from_value(value) else {
-        return f64::from_bits(crate::value::TAG_UNDEFINED);
-    };
-    let key = named_key(name);
-    js_object_get_field_by_name_f64(obj, key)
-}
-
-fn is_callable_value(value: f64) -> bool {
-    let raw = raw_ptr_from_value(value);
-    if raw < 0x10000 || crate::buffer::is_registered_buffer(raw) {
-        return false;
-    }
-    unsafe {
-        gc_type_for_ptr(raw) == Some(crate::gc::GC_TYPE_CLOSURE)
-            && crate::closure::is_closure_ptr(raw)
-    }
-}
-
-fn has_named_callable(value: f64, name: &[u8]) -> bool {
-    is_callable_value(get_named_value(value, name))
-}
-
-fn invalid_chunk_error(value: f64) -> f64 {
-    let jsval = JSValue::from_bits(value.to_bits());
-    let (kind, detail) = if jsval.is_int32() {
-        ("number", format!(" ({})", jsval.as_int32()))
-    } else if jsval.is_number() {
-        ("number", format!(" ({})", value))
-    } else if jsval.is_bool() {
-        ("boolean", String::new())
-    } else if jsval.is_undefined() {
-        ("undefined", String::new())
-    } else if jsval.is_null() {
-        ("null", String::new())
-    } else if is_callable_value(value) {
-        ("function", String::new())
-    } else {
-        ("object", String::new())
-    };
-    let msg = format!(
-        "The \"chunk\" argument must be of type string or an instance of Buffer, TypedArray, or DataView. Received type {}{}",
-        kind, detail
-    );
-    let msg_ptr = js_string_from_bytes(msg.as_ptr(), msg.len() as u32);
-    crate::node_submodules::register_error_code_pub(msg_ptr, "ERR_INVALID_ARG_TYPE");
-    let err = crate::error::js_typeerror_new(msg_ptr);
-    boxed_pointer(err as *const u8)
-}
-
-fn append_string_value_bytes(value: f64, out: &mut Vec<u8>) {
-    let ptr = crate::value::js_get_string_pointer_unified(value) as *const StringHeader;
-    append_string_ptr_bytes(ptr, out);
-}
-
-fn append_string_ptr_bytes(ptr: *const StringHeader, out: &mut Vec<u8>) {
-    if ptr.is_null() || (ptr as usize) < 0x10000 {
-        return;
-    }
-    unsafe {
-        let len = (*ptr).byte_len as usize;
-        let data = (ptr as *const u8).add(std::mem::size_of::<StringHeader>());
-        out.extend_from_slice(std::slice::from_raw_parts(data, len));
-    }
-}
-
-fn append_buffer_value_bytes(raw: usize, out: &mut Vec<u8>) {
-    if raw < 0x10000 || !crate::buffer::is_registered_buffer(raw) {
-        return;
-    }
-    unsafe {
-        let buf = raw as *const crate::buffer::BufferHeader;
-        let len = (*buf).length as usize;
-        let data = crate::buffer::buffer_data(buf);
-        out.extend_from_slice(std::slice::from_raw_parts(data, len));
-    }
-}
-
-fn append_number_chunk(value: f64, jsval: JSValue, out: &mut Vec<u8>) {
-    let text = if jsval.is_int32() {
-        jsval.as_int32().to_string()
-    } else if value.is_finite() && value.fract() == 0.0 {
-        (value as i64).to_string()
-    } else {
-        value.to_string()
-    };
-    out.extend_from_slice(text.as_bytes());
-}
-
-fn append_array_chunk_bytes(
-    raw: usize,
-    out: &mut Vec<u8>,
-    mode: ChunkMode,
-    depth: u8,
-) -> Result<(), f64> {
-    if raw < 0x10000 {
-        return Ok(());
-    }
-    let arr = raw as *const crate::array::ArrayHeader;
-    let len = crate::array::js_array_length(arr);
-    for i in 0..len {
-        let chunk = crate::array::js_array_get_f64(arr, i);
-        append_chunk_bytes_for_consumer(chunk, out, mode, depth + 1)?;
-    }
-    Ok(())
-}
-
-fn append_chunk_bytes_for_consumer(
-    value: f64,
-    out: &mut Vec<u8>,
-    mode: ChunkMode,
-    depth: u8,
-) -> Result<(), f64> {
-    if depth > 16 {
-        return Err(invalid_chunk_error(value));
-    }
-    let jsval = JSValue::from_bits(value.to_bits());
-    if jsval.is_any_string() {
-        append_string_value_bytes(value, out);
-        return Ok(());
-    }
-    if jsval.is_int32() || (jsval.is_number() && value.is_finite()) {
-        if mode == ChunkMode::Text {
-            return Err(invalid_chunk_error(value));
-        }
-        append_number_chunk(value, jsval, out);
-        return Ok(());
-    }
-
-    let raw = raw_ptr_from_value(value);
-    if raw < 0x10000 {
-        return Err(invalid_chunk_error(value));
-    }
-    if crate::buffer::is_registered_buffer(raw) {
-        append_buffer_value_bytes(raw, out);
-        return Ok(());
-    }
-
-    unsafe {
-        match gc_type_for_ptr(raw) {
-            Some(crate::gc::GC_TYPE_ARRAY | crate::gc::GC_TYPE_LAZY_ARRAY) => {
-                append_array_chunk_bytes(raw, out, mode, depth)
-            }
-            Some(crate::gc::GC_TYPE_OBJECT) => {
-                if let Some(Ok(chunks)) =
-                    crate::node_stream::js_node_stream_collect_chunks_result(value)
-                {
-                    append_chunk_bytes_for_consumer(chunks, out, mode, depth + 1)
-                } else {
-                    Err(invalid_chunk_error(value))
-                }
-            }
-            Some(crate::gc::GC_TYPE_STRING) => {
-                append_string_ptr_bytes(raw as *const StringHeader, out);
-                Ok(())
-            }
-            _ => Err(invalid_chunk_error(value)),
-        }
-    }
-}
-
-fn chunks_to_bytes(chunks: f64, mode: ChunkMode) -> Result<Vec<u8>, f64> {
-    let mut out = Vec::new();
-    append_chunk_bytes_for_consumer(chunks, &mut out, mode, 0)?;
-    Ok(out)
-}
-
-fn finish_consumer_from_chunks(kind: ConsumerKind, chunks: f64) -> Result<f64, f64> {
-    let bytes = chunks_to_bytes(chunks, kind.chunk_mode())?;
-    match kind {
-        ConsumerKind::Text => Ok(bytes_to_text_value(&bytes)),
-        ConsumerKind::Json => {
-            let text = bytes_to_text_value(&bytes);
-            let text_ptr = crate::value::js_get_string_pointer_unified(text) as *const StringHeader;
-            unsafe { crate::json::js_json_parse_result(text_ptr).map(|v| f64::from_bits(v.bits())) }
-        }
-        ConsumerKind::Buffer => Ok(bytes_to_buffer_value(&bytes)),
-        ConsumerKind::ArrayBuffer => Ok(bytes_to_array_buffer_value(&bytes)),
-        ConsumerKind::Bytes => Ok(bytes_to_uint8_array_value(&bytes)),
-        ConsumerKind::Blob => Ok(blob_value_from_bytes(&bytes)),
-    }
-}
-
-fn promise_from_consumer_chunks(kind: ConsumerKind, chunks: Result<f64, f64>) -> f64 {
-    match chunks.and_then(|chunks| finish_consumer_from_chunks(kind, chunks)) {
-        Ok(value) => promise_value(value),
-        Err(err) => promise_rejected(err),
-    }
-}
-
-fn settle_consumer_from_chunks(promise: *mut crate::Promise, kind: ConsumerKind, chunks: f64) {
-    if promise.is_null() {
-        return;
-    }
-    match finish_consumer_from_chunks(kind, chunks) {
-        Ok(value) => crate::promise::js_promise_resolve(promise, value),
-        Err(err) => crate::promise::js_promise_reject(promise, err),
-    }
-}
-
-fn promise_ptr_from_value(value: f64) -> Option<*mut crate::Promise> {
-    if crate::promise::js_value_is_promise(value) == 0 {
-        return None;
-    }
-    let raw = raw_ptr_from_value(value);
-    if raw < 0x10000 {
-        None
-    } else {
-        Some(raw as *mut crate::Promise)
-    }
-}
-
-fn registered_readable_stream_reader(stream: f64) -> Option<f64> {
-    if !is_integral_handle_value(stream) {
-        return None;
-    }
-    let f = STREAM_CONSUMER_GET_READER.load(Ordering::Acquire);
-    if f.is_null() {
-        return None;
-    }
-    let reader = unsafe {
-        let func: StreamGetReaderFn = std::mem::transmute(f);
-        func(stream)
-    };
-    if is_undefined_value(reader) {
-        None
-    } else {
-        Some(reader)
-    }
-}
-
-fn registered_reader_read_promise(reader: f64) -> Option<*mut crate::Promise> {
-    if !is_integral_handle_value(reader) {
-        return None;
-    }
-    let f = STREAM_CONSUMER_READER_READ.load(Ordering::Acquire);
-    if f.is_null() {
-        return None;
-    }
-    Some(unsafe {
-        let func: StreamReaderReadFn = std::mem::transmute(f);
-        func(reader)
-    })
-}
-
-fn call_collector_method(
-    receiver: f64,
-    method: CollectMethod,
-    step: *const ClosureHeader,
-    reject: *const ClosureHeader,
-) {
-    if let CollectMethod::Read = method {
-        if let Some(promise) = registered_reader_read_promise(receiver) {
-            crate::promise::js_promise_then(promise, step, reject);
-            return;
-        }
-    }
-
-    let name = method.name();
-    let result = unsafe {
-        crate::object::js_native_call_method(
-            receiver,
-            name.as_ptr() as *const i8,
-            name.len(),
-            std::ptr::null(),
-            0,
-        )
-    };
-    if let Some(promise) = promise_ptr_from_value(result) {
-        crate::promise::js_promise_then(promise, step, reject);
-    } else {
-        consumer_collect_step(step, result);
-    }
-}
-
-fn collect_by_method_promise(kind: ConsumerKind, receiver: f64, method: CollectMethod) -> f64 {
-    let result_promise = crate::promise::js_promise_new();
-    let result_arr = crate::array::js_array_alloc(0);
-    let step = js_closure_alloc(consumer_collect_step as *const u8, 6);
-    let reject = js_closure_alloc(consumer_collect_rejected as *const u8, 1);
-    js_closure_set_capture_ptr(step, 0, result_promise as i64);
-    js_closure_set_capture_ptr(step, 1, result_arr as i64);
-    js_closure_set_capture_f64(step, 2, receiver);
-    js_closure_set_capture_ptr(step, 3, reject as i64);
-    js_closure_set_capture_ptr(step, 4, method as i64);
-    js_closure_set_capture_ptr(step, 5, kind as i64);
-    js_closure_set_capture_ptr(reject, 0, result_promise as i64);
-    call_collector_method(receiver, method, step, reject);
-    boxed_pointer(result_promise as *const u8)
-}
-
-fn call_symbol_async_iterator(stream: f64) -> Option<f64> {
-    let sym = crate::symbol::well_known_symbol("asyncIterator");
-    if sym.is_null() {
-        return None;
-    }
-    let sym_f64 = boxed_pointer(sym as *const u8);
-    let method = unsafe { crate::symbol::js_object_get_symbol_property(stream, sym_f64) };
-    if !is_callable_value(method) {
-        return None;
-    }
-    let prev_this = crate::object::js_implicit_this_set(stream);
-    let iterator = unsafe { crate::closure::js_native_call_value(method, std::ptr::null(), 0) };
-    crate::object::js_implicit_this_set(prev_this);
-    if iterator.to_bits() == crate::value::TAG_UNDEFINED {
-        None
-    } else {
-        Some(iterator)
-    }
-}
-
-fn async_consumer_promise(kind: ConsumerKind, stream: f64) -> Option<f64> {
-    if let Some(iterator) = call_symbol_async_iterator(stream) {
-        if has_named_callable(iterator, b"next") {
-            return Some(collect_by_method_promise(
-                kind,
-                iterator,
-                CollectMethod::Next,
-            ));
-        }
-    }
-    if has_named_callable(stream, b"next") {
-        return Some(collect_by_method_promise(kind, stream, CollectMethod::Next));
-    }
-    if has_named_callable(stream, b"getReader") {
-        let reader = unsafe {
-            crate::object::js_native_call_method(
-                stream,
-                b"getReader".as_ptr() as *const i8,
-                b"getReader".len(),
-                std::ptr::null(),
-                0,
-            )
-        };
-        if reader.to_bits() == crate::value::TAG_UNDEFINED {
-            return Some(promise_rejected(invalid_chunk_error(stream)));
-        }
-        return Some(collect_by_method_promise(kind, reader, CollectMethod::Read));
-    }
-    if let Some(reader) = registered_readable_stream_reader(stream) {
-        return Some(collect_by_method_promise(kind, reader, CollectMethod::Read));
-    }
-    None
-}
-
-fn consume_stream(kind: ConsumerKind, stream: f64) -> f64 {
-    if let Some(chunks) = crate::node_stream::js_node_stream_collect_chunks_result(stream) {
-        return promise_from_consumer_chunks(kind, chunks);
-    }
-    if let Some(promise) = async_consumer_promise(kind, stream) {
-        return promise;
-    }
-    let empty = crate::array::js_array_alloc(0);
-    promise_from_consumer_chunks(kind, Ok(boxed_pointer(empty as *const u8)))
-}
-
-extern "C" fn consumer_collect_rejected(closure: *const ClosureHeader, reason: f64) -> f64 {
-    let promise = js_closure_get_capture_ptr(closure, 0) as *mut crate::Promise;
-    crate::promise::js_promise_reject(promise, reason);
-    0.0
-}
-
-extern "C" fn consumer_collect_step(closure: *const ClosureHeader, iter_result: f64) -> f64 {
-    let promise = js_closure_get_capture_ptr(closure, 0) as *mut crate::Promise;
-    let mut result_arr = js_closure_get_capture_ptr(closure, 1) as *mut crate::array::ArrayHeader;
-    let receiver = js_closure_get_capture_f64(closure, 2);
-    let reject = js_closure_get_capture_ptr(closure, 3) as *const ClosureHeader;
-    let method = CollectMethod::from_i64(js_closure_get_capture_ptr(closure, 4));
-    let kind = ConsumerKind::from_i64(js_closure_get_capture_ptr(closure, 5));
-    if promise.is_null() || result_arr.is_null() {
-        return 0.0;
-    }
-
-    let Some(result_obj) = object_ptr_from_value(iter_result) else {
-        let arr_value = boxed_pointer(result_arr as *const u8);
-        settle_consumer_from_chunks(promise, kind, arr_value);
-        return 0.0;
-    };
-
-    let done = js_object_get_field_by_name_f64(result_obj, named_key(b"done"));
-    if crate::value::js_is_truthy(done) != 0 {
-        let arr_value = boxed_pointer(result_arr as *const u8);
-        settle_consumer_from_chunks(promise, kind, arr_value);
-        return 0.0;
-    }
-
-    let value = js_object_get_field_by_name_f64(result_obj, named_key(b"value"));
-    result_arr = crate::array::js_array_push_f64(result_arr, value);
-    js_closure_set_capture_ptr(closure as *mut ClosureHeader, 1, result_arr as i64);
-    call_collector_method(receiver, method, closure, reject);
-    0.0
-}
-
-const CLASS_ID_BLOB: u32 = 0xFFFF0026;
-
-extern "C" fn blob_text_method(closure: *const ClosureHeader) -> f64 {
-    let bytes = captured_blob_bytes(closure);
-    promise_value(bytes_to_text_value(&bytes))
-}
-
-extern "C" fn blob_array_buffer_method(closure: *const ClosureHeader) -> f64 {
-    let bytes = captured_blob_bytes(closure);
-    promise_value(bytes_to_array_buffer_value(&bytes))
-}
-
-extern "C" fn blob_bytes_method(closure: *const ClosureHeader) -> f64 {
-    let bytes = captured_blob_bytes(closure);
-    promise_value(bytes_to_uint8_array_value(&bytes))
-}
-
-extern "C" fn blob_slice_method(
-    closure: *const ClosureHeader,
-    start: f64,
-    end: f64,
-    content_type: f64,
-) -> f64 {
-    let bytes = captured_blob_bytes(closure);
-    let len = bytes.len() as i64;
-    let normalize = |value: f64, default: i64| -> i64 {
-        if value.is_nan() || value.to_bits() == crate::value::TAG_UNDEFINED {
-            return default;
-        }
-        let n = value as i64;
-        if n < 0 {
-            (len + n).max(0)
-        } else {
-            n.min(len)
-        }
-    };
-    let lo = normalize(start, 0);
-    let hi = normalize(end, len);
-    let (lo, hi) = if hi < lo { (lo, lo) } else { (lo, hi) };
-    let content_type = string_from_value(content_type).unwrap_or_default();
-    blob_value_from_bytes_and_type(&bytes[lo as usize..hi as usize], &content_type)
-}
-
-extern "C" fn blob_stream_method(closure: *const ClosureHeader) -> f64 {
-    let bytes = captured_blob_bytes(closure);
-    crate::node_stream::js_node_stream_readable_from(bytes_to_uint8_array_value(&bytes))
-}
-
-fn captured_blob_bytes(closure: *const ClosureHeader) -> Vec<u8> {
-    let raw = js_closure_get_capture_ptr(closure, 0) as usize;
-    if raw < 0x10000 || !crate::buffer::is_registered_buffer(raw) {
-        return Vec::new();
-    }
-    unsafe {
-        let buf = raw as *const crate::buffer::BufferHeader;
-        let len = (*buf).length as usize;
-        let data = crate::buffer::buffer_data(buf);
-        std::slice::from_raw_parts(data, len).to_vec()
-    }
-}
-
-fn set_named_value(obj: *mut ObjectHeader, name: &[u8], value: f64) {
-    let key = js_string_from_bytes(name.as_ptr(), name.len() as u32);
-    js_object_set_field_by_name(obj, key, value);
-}
-
-#[allow(clippy::missing_transmute_annotations)]
-fn blob_method_value(
-    func: *const u8,
-    arity: u32,
-    backing: *mut crate::buffer::BufferHeader,
-) -> f64 {
-    js_register_closure_arity(func, arity);
-    let closure = js_closure_alloc(func, 1);
-    js_closure_set_capture_ptr(closure, 0, backing as i64);
-    f64::from_bits(JSValue::pointer(closure as *const u8).bits())
-}
-
-fn blob_value_from_bytes(bytes: &[u8]) -> f64 {
-    blob_value_from_bytes_and_type(bytes, "")
-}
-
-fn blob_value_from_bytes_and_type(bytes: &[u8], content_type: &str) -> f64 {
-    let backing = buffer_from_bytes(bytes, false, false);
-    let obj = js_object_alloc(CLASS_ID_BLOB, 7);
-    set_named_value(obj, b"size", bytes.len() as f64);
-    set_named_value(obj, b"type", bytes_to_text_value(content_type.as_bytes()));
-    set_named_value(
-        obj,
-        b"text",
-        blob_method_value(blob_text_method as *const u8, 0, backing),
-    );
-    set_named_value(
-        obj,
-        b"arrayBuffer",
-        blob_method_value(blob_array_buffer_method as *const u8, 0, backing),
-    );
-    set_named_value(
-        obj,
-        b"bytes",
-        blob_method_value(blob_bytes_method as *const u8, 0, backing),
-    );
-    set_named_value(
-        obj,
-        b"slice",
-        blob_method_value(blob_slice_method as *const u8, 3, backing),
-    );
-    set_named_value(
-        obj,
-        b"stream",
-        blob_method_value(blob_stream_method as *const u8, 0, backing),
-    );
-    f64::from_bits(JSValue::pointer(obj as *const u8).bits())
-}
-
-fn string_from_value(value: f64) -> Option<String> {
-    let jsval = JSValue::from_bits(value.to_bits());
-    if !jsval.is_any_string() {
-        return None;
-    }
-    let ptr = crate::value::js_get_string_pointer_unified(value) as *const StringHeader;
-    if ptr.is_null() || (ptr as usize) < 0x10000 {
-        return None;
-    }
-    unsafe {
-        let len = (*ptr).byte_len as usize;
-        let data = (ptr as *const u8).add(std::mem::size_of::<StringHeader>());
-        Some(String::from_utf8_lossy(std::slice::from_raw_parts(data, len)).into_owned())
-    }
-}
-
-extern "C" fn thunk_consumers_text(_closure: *const ClosureHeader, stream: f64) -> f64 {
-    consume_stream(ConsumerKind::Text, stream)
-}
-
-extern "C" fn thunk_consumers_json(_closure: *const ClosureHeader, stream: f64) -> f64 {
-    consume_stream(ConsumerKind::Json, stream)
-}
-
-extern "C" fn thunk_consumers_buffer(_closure: *const ClosureHeader, stream: f64) -> f64 {
-    consume_stream(ConsumerKind::Buffer, stream)
-}
-
-extern "C" fn thunk_consumers_arrayBuffer(_closure: *const ClosureHeader, stream: f64) -> f64 {
-    consume_stream(ConsumerKind::ArrayBuffer, stream)
-}
-
-extern "C" fn thunk_consumers_bytes(_closure: *const ClosureHeader, stream: f64) -> f64 {
-    consume_stream(ConsumerKind::Bytes, stream)
-}
-
-extern "C" fn thunk_consumers_blob(_closure: *const ClosureHeader, stream: f64) -> f64 {
-    consume_stream(ConsumerKind::Blob, stream)
-}
+mod blob;
+mod consumers;
+mod fs_promises;
+mod stream_promises;
+mod timers;
+
+use consumers::{
+    thunk_consumers_arrayBuffer, thunk_consumers_blob, thunk_consumers_buffer,
+    thunk_consumers_bytes, thunk_consumers_json, thunk_consumers_text,
+};
+// Re-export at the `node_submodules` root so perry-stdlib's
+// `perry_runtime::node_submodules::js_register_stream_consumer_callbacks`
+// call site keeps resolving after the consumers split.
+pub use consumers::js_register_stream_consumer_callbacks;
+use fs_promises::{
+    thunk_fs_promises_access, thunk_fs_promises_appendFile, thunk_fs_promises_chmod,
+    thunk_fs_promises_chown, thunk_fs_promises_copyFile, thunk_fs_promises_cp,
+    thunk_fs_promises_glob, thunk_fs_promises_lchown, thunk_fs_promises_link,
+    thunk_fs_promises_lstat, thunk_fs_promises_lutimes, thunk_fs_promises_mkdir,
+    thunk_fs_promises_mkdtemp, thunk_fs_promises_open, thunk_fs_promises_opendir,
+    thunk_fs_promises_readFile, thunk_fs_promises_readdir, thunk_fs_promises_readlink,
+    thunk_fs_promises_realpath, thunk_fs_promises_rename, thunk_fs_promises_rm,
+    thunk_fs_promises_rmdir, thunk_fs_promises_stat, thunk_fs_promises_statfs,
+    thunk_fs_promises_symlink, thunk_fs_promises_truncate, thunk_fs_promises_unlink,
+    thunk_fs_promises_utimes, thunk_fs_promises_watch, thunk_fs_promises_writeFile,
+    thunk_readline_Interface, thunk_readline_Readline, thunk_readline_createInterface,
+};
+use stream_promises::{thunk_streamP_finished, thunk_streamP_pipeline, value_from_ptr};
+use timers::{
+    thunk_timers_scheduler, thunk_timers_setInterval, timers_ns_clear_immediate,
+    timers_ns_clear_interval, timers_ns_clear_timeout, timers_ns_set_immediate,
+    timers_ns_set_interval, timers_ns_set_timeout, timers_promises_set_immediate,
+    timers_promises_set_timeout,
+};
 
 // node:sys is a deprecated alias for node:util — point each export at
 // the same thunks until util's named-export surface is wired up. The
@@ -1975,6 +735,10 @@ pub unsafe extern "C" fn js_node_submodule_namespace(
 
 #[cfg(test)]
 mod tests {
+    use super::blob::string_from_value;
+    use super::stream_promises::{
+        abort_error_value, get_object_property, object_ptr_from_value, undefined_value,
+    };
     use super::*;
     use std::cell::RefCell;
